Move getSidebarPreferences to selectors/entities/preferences
